fix: mudança de tipo de algumas coisas

+++ header/ambient_light.hpp
@@ -7,7 +7,7 @@ class AmbientLight : public Light {
     public:
         AmbientLight(Color intensity);
     
-        Color calculate_intensity(Vector P, Vector N, Vector V, int s,  Object* o, bool has_shadow);
+        Color calculate_intensity(Vector P, Vector N, Vector V, double s,  Object* o, bool has_shadow);
         Vector get_l(Vector P);
 };
